+++ src/components/MyPageInfo.js
@@ -24,11 +24,11 @@ import Spinner from "../shared/Spinner";
 import DeleteUserModal from "../modal/DeleteUserModal";
 
 // MyPageInfo의 함수형 컴포넌트를 만든다.
-const MyPageInfo = (props) => {
+const MyPageInfo = props => {
   const dispatch = useDispatch();
   // const userId = useSelector((state) => state.user.userId);
   const userId = props.match.params.id;
-  const myUserId = useSelector((state) => state.user.userId);
+  const myUserId = useSelector(state => state.user.userId);
   // console.log(props);
 
   const [filter, setFilter] = React.useState("소개");
@@ -69,7 +69,7 @@ const MyPageInfo = (props) => {
   }, [filter, editMyProfile]);
 
   React.useEffect(
-    (e) => {
+    e => {
       if (currentClick !== null) {
         let current = document.getElementById(currentClick);
         current.style.color = "#333";
@@ -85,7 +85,7 @@ const MyPageInfo = (props) => {
     },
     [currentClick]
   );
-  const GetClick = (e) => {
+  const GetClick = e => {
     setCurrentClick(e);
     // console.log(e);
   };
@@ -481,7 +481,7 @@ const MyPageInfo = (props) => {
                             padding: "7px",
                           }}
                           defaultValue={mydata.user.nickname}
-                          onChange={(e) => {
+                          onChange={e => {
                             setNickName(e.target.value);
                           }}
                         ></input>
@@ -515,7 +515,7 @@ const MyPageInfo = (props) => {
                             padding: "7px",
                           }}
                           defaultValue={mydata.user.email}
-                          onChange={(e) => {
+                          onChange={e => {
                             setEmail(e.target.value);
                           }}
                         ></input>
@@ -539,7 +539,7 @@ const MyPageInfo = (props) => {
                           styles={styles}
                           className="basic-multi-select"
                           classNamePrefix="select"
-                          onChange={(e) => {
+                          onChange={e => {
                             let techStack = [];
                             let arr = e;
                             let idx = 0;
@@ -973,7 +973,7 @@ const MyPageInfo = (props) => {
               >
                 <Filter
                   id="recruitment"
-                  onClick={(e) => {
+                  onClick={e => {
                     setFilter("모집");
                     GetClick(e.target.id);
                   }}
@@ -982,7 +982,7 @@ const MyPageInfo = (props) => {
                 </Filter>
                 <Filter
                   id="progress"
-                  onClick={(e) => {
+                  onClick={e => {
                     setFilter("진행중");
                     GetClick(e.target.id);
                   }}
@@ -991,7 +991,7 @@ const MyPageInfo = (props) => {
                 </Filter>
                 <Filter
                   id="bookmark"
-                  onClick={(e) => {
+                  onClick={e => {
                     setFilter("관심");
                     GetClick(e.target.id);
                   }}
@@ -1000,7 +1000,7 @@ const MyPageInfo = (props) => {
                 </Filter>
                 <Filter
                   id="finish"
-                  onClick={(e) => {
+                  onClick={e => {
                     setFilter("완료");
                     GetClick(e.target.id);
                   }}
@@ -1009,7 +1009,7 @@ const MyPageInfo = (props) => {
                 </Filter>
                 <Filter
                   id="introduction"
-                  onClick={(e) => {
+                  onClick={e => {
                     setFilter("소개");
                     GetClick(e.target.id);
                   }}
@@ -1059,7 +1059,10 @@ const MyPageInfo = (props) => {
                 )}
               </Grid>
               {filter === "완료" && (
-                <MypagePostList {...endProject}></MypagePostList>
+                <MypagePostList
+                  filter={filter}
+                  {...endProject}
+                ></MypagePostList>
               )}
               <Grid margin="0 0 0 34%" width="49%">
                 {filter === "완료" && endProject.length === 0 && (
