+++ src/components/applyUserModal/ApplyIndex.js
@@ -2,5 +2,6 @@ import Apply from "./Apply";
 import Cancel from "./Cancel";
 import TeamExit from "./TeamExit";
 import SubmitUrl from "./SubmitUrl";
+import Liked from "./Liked";
 
-export { Apply, Cancel, TeamExit, SubmitUrl };
+export { Apply, Cancel, TeamExit, SubmitUrl, Liked };
